add error boundary component
