+++ .github/workflows/codesanbox-link.yml
@@ -1,6 +1,8 @@
 name: CodeSandbox Link
 
-on: pull_request
+on:
+  pull_request:
+    types: [opened, reopened]
 
 jobs:
   add-codesandbox-link-comment:
