+++ package.json
@@ -66,17 +66,17 @@
     "@types/jsdom": "^21.1.1",
     "@types/node": "^18.15.11",
     "@types/prettier": "^2.7.2",
-    "@types/react": "^18.0.32",
+    "@types/react": "^18.0.33",
     "@types/react-dom": "^18.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.57.0",
-    "@typescript-eslint/parser": "^5.57.0",
+    "@typescript-eslint/eslint-plugin": "^5.57.1",
+    "@typescript-eslint/parser": "^5.57.1",
     "babel-loader": "^9.1.2",
-    "chromatic": "^6.17.2",
+    "chromatic": "^6.17.3",
     "copy-webpack-plugin": "^11.0.0",
-    "core-js": "^3.29.1",
+    "core-js": "^3.30.0",
     "css-loader": "^6.7.3",
     "css-minimizer-webpack-plugin": "^5.0.0",
-    "eslint": "^8.37.0",
+    "eslint": "^8.38.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.27.5",
@@ -94,14 +94,14 @@
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
     "jest-extended": "^3.2.4",
-    "lint-staged": "^13.2.0",
+    "lint-staged": "^13.2.1",
     "loki": "^0.31.1",
     "mini-css-extract-plugin": "^2.7.5",
-    "npm-check-updates": "^16.10.0",
+    "npm-check-updates": "^16.10.7",
     "prettier": "^2.8.7",
     "prettier-plugin-packagejson": "^2.4.3",
     "prettier-plugin-sh": "^0.12.8",
-    "sass": "^1.60.0",
+    "sass": "^1.61.0",
     "sass-loader": "^13.2.2",
     "style-loader": "^3.3.2",
     "stylelint": "^15.4.0",
@@ -110,7 +110,7 @@
     "ts-loader": "^9.4.2",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.5",
-    "webpack": "^5.77.0",
+    "webpack": "^5.78.0",
     "webpack-bundle-analyzer": "^4.8.0",
     "webpack-cli": "^5.0.1",
     "webpack-dev-server": "^4.13.2"
