+++ src/components/Header/index.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { Header } from ".";
-import { text } from "../../constants";
+import { text } from "../../helpers/constants";
 
 describe("Header", () => {
   it("should render header", () => {
