+++ src/components/Header/index.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { HeaderProps } from "../../types";
+import { HeaderProps } from "../../types/types";
 
 export const Header: FC<HeaderProps> = ({ level, text }) => {
   const Tag = `h${level}` as `h${typeof level}`;
