+++ src/types/declaration.d.ts
