+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
   "include": ["./src/**/*", "./test/**/*"],
   "compilerOptions": {
-    "target": "ES2016",
+    "target": "ESNext",
     "jsx": "react-jsx",
     "module": "CommonJS",
     "outDir": "./out",
@@ -10,5 +10,6 @@
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "skipLibCheck": true
-  }
+  },
+  "$schema": "https://json.schemastore.org/tsconfig"
 }
